+++ config/initializers/notifications.rb
@@ -1,0 +1,1 @@
+Notification.uri = Rails.my_config(:notifications_uri)